replaced script with entry point in setup.py
The `setup.py` specified previously a script which caused problems when
running pyicontract-lint on Windows (*e.g.*, Windows could not find the
interpreter since the shebang is ignored). This change removes the
script and specifies an entry point in `setup.py` so that the executable
is correctly created on Windows upon installation.

See also:
https://setuptools.readthedocs.io/en/latest/setuptools.html#automatic-script-creation