handled paths platform-depedentent
We erroneously assumed the platform to be POSIX. Consequently, the paths
and temporary files were not handled properly (*e.g.*, `as_posix()` was
used, `NamedTemporaryFile` was re-opened, the default interpreter was
assumed to be `python3` *etc*.).

This change fixes all these issues and the precommit checks run
successfully on Windows.