added output on no errors if verbose (#25)
Previously, verbose output was empty if there were no errors which was
confusing for the users. This change adds a message to the verbose
output signaling that no errors were detected.

Additionally, all new lines were made platform-dependent in the output
while the previous line separator was set to Linux `\n`.

This change fixes #23.