Added support for astroid ≥ 2.6.0
The code in astroid underwent a minor refactoring and broke the public
interface (`ALL_NODE_CLASSES` were moved to `astroid.nodes` module).
This had repercussions on pyicontract-lint since an assertion depends on
it.

This patch examines the version of astroid and adapts the assertion
according to the version so that both older and newer versions of
astroid are supported.

Fixes #37.